Work in latest Gophercloud
This commit includes a boat load of changes.  We now have attach actions
merged in Gophercloud, so we went through and added all the necessary
items to make that work.

Also, this removes a bunch of the hard coding that was present just to
get a proof of concept going.  Currently this is functional with the
default volume-type in Cinder (tested on LVM and SolidFire).

There are a number of annoying things that the Docker Volume API does
around error recovery as well as persisting attachment/mount info,
in other words, there's still a bit of work needed to clean up and
make things more robust.

Code is a work in progress, but currently you can create, rm, ls
etc to Cinder from the Volume API (Docker CLI), and you can also
do a docker run, create/attach all at once:
    `docker run -ti -v foo:/data --volume-driver=cinder ubuntu bash`