Add a README and install script
Add a README with a least a little info on how one might install
and use this.  Also include a simple install script that will pull
the release binary and install it.

I've also added UUID detection so we don't have to make something
up for the Node any more.  Now, the driver will just try and
determine/use the UUID of the root disk of the node.  Of course
if one chooses to specify this in the config file that's fine too.

I've also added some interface/ip detection so those aren't required
in the config file any longer (of course they can still be specified
if desired, and you should desire in normal cases).

So we now have a minimal config to get up and running and test easily.

Most of this was inspired by a need to deploy things quickly and in
an automated fashion to docker-machine nodes.