Add a version number
We didn't have any versioning exposed to the user which is less
than useful.

Add a version string printout at startup, as well as an option
to run with "--version" which will just print out the version and
exit.

We'll go and bump to 0.11 now and tag it as a new release.